+++ src/solver.cpp
@@ -91,6 +91,7 @@ simple_gmres_euler(const P dt, kronmult_matrix<P> const &mat,
       restart, max_iter, tolerance);
 }
 
+#ifdef KRON_MODE_GLOBAL
 template<typename P, resource resrc>
 gmres_info<P>
 simple_gmres_euler(const P dt, matrix_entry mentry,
@@ -110,6 +111,7 @@ simple_gmres_euler(const P dt, matrix_entry mentry,
       fk::vector<P, mem_type::view, resrc>(x), b, no_op_preconditioner<P>(),
       restart, max_iter, tolerance);
 }
+#endif
 
 /*! Generates a default number inner iterations when no use input is given
  * \param num_cols Number of columns in the A matrix.
@@ -406,12 +408,14 @@ simple_gmres_euler(const double dt, kronmult_matrix<double> const &mat,
                    int const restart, int const max_iter,
                    double const tolerance);
 
+#ifdef KRON_MODE_GLOBAL
 template gmres_info<double>
 simple_gmres_euler(const double dt, matrix_entry mentry,
                    global_kron_matrix<double> const &mat,
                    fk::vector<double, mem_type::owner, resource::host> &x,
                    fk::vector<double, mem_type::owner, resource::host> const &b,
                    int const restart, int const max_iter, double const tolerance);
+#endif
 
 template int default_gmres_restarts<double>(int num_cols);
 
@@ -448,12 +452,14 @@ simple_gmres_euler(const float dt, kronmult_matrix<float> const &mat,
                    int const restart, int const max_iter,
                    float const tolerance);
 
+#ifdef KRON_MODE_GLOBAL
 template gmres_info<float>
 simple_gmres_euler(const float dt, matrix_entry mentry,
                    global_kron_matrix<float> const &mat,
                    fk::vector<float, mem_type::owner, resource::host> &x,
                    fk::vector<float, mem_type::owner, resource::host> const &b,
                    int const restart, int const max_iter, float const tolerance);
+#endif
 
 template int default_gmres_restarts<float>(int num_cols);
